+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/Automator.kt
@@ -156,10 +156,10 @@ class Automator private constructor() {
         return getWindowTrees(uiDevice, uiAutomation)
     }
 
-    fun tap(uiSelector: UiSelector, bySelector: BySelector, index: Int) {
+    fun tap(uiSelector: UiSelector, bySelector: BySelector, index: Int, timeout: Long? = null) {
         Logger.d("tap(): $uiSelector, $bySelector")
 
-        if (waitForView(bySelector, index) == null) {
+        if (waitForView(bySelector, index, timeout) == null) {
             throw UiObjectNotFoundException("$uiSelector")
         }
 
@@ -169,12 +169,12 @@ class Automator private constructor() {
         delay()
     }
 
-    fun doubleTap(uiSelector: UiSelector, bySelector: BySelector, index: Int) {
+    fun doubleTap(uiSelector: UiSelector, bySelector: BySelector, index: Int, timeout: Long? = null) {
         Logger.d("doubleTap(): $uiSelector, $bySelector")
 
         val uiObject = uiDevice.findObject(uiSelector)
 
-        if (waitForView(bySelector, index) == null) {
+        if (waitForView(bySelector, index, timeout) == null) {
             throw UiObjectNotFoundException("$uiSelector")
         }
 
@@ -187,11 +187,11 @@ class Automator private constructor() {
         delay()
     }
 
-    fun enterText(text: String, index: Int, keyboardBehavior: KeyboardBehavior) {
+    fun enterText(text: String, index: Int, keyboardBehavior: KeyboardBehavior, timeout: Long? = null) {
         Logger.d("enterText(text: $text, index: $index)")
 
         val selector = By.clazz(EditText::class.java)
-        if (waitForView(selector, index) == null) {
+        if (waitForView(selector, index, timeout) == null) {
             throw UiObjectNotFoundException("$selector")
         }
 
@@ -216,11 +216,12 @@ class Automator private constructor() {
         uiSelector: UiSelector,
         bySelector: BySelector,
         index: Int,
-        keyboardBehavior: KeyboardBehavior
+        keyboardBehavior: KeyboardBehavior,
+        timeout: Long? = null
     ) {
         Logger.d("enterText($text): $uiSelector, $bySelector")
 
-        if (waitForView(bySelector, index) == null) {
+        if (waitForView(bySelector, index, timeout) == null) {
             throw UiObjectNotFoundException("$uiSelector")
         }
 
@@ -269,10 +270,10 @@ class Automator private constructor() {
         delay()
     }
 
-    fun waitUntilVisible(uiSelector: UiSelector, bySelector: BySelector, index: Int) {
+    fun waitUntilVisible(uiSelector: UiSelector, bySelector: BySelector, index: Int, timeout: Long? = null) {
         Logger.d("waitUntilVisible(): $uiSelector, $bySelector")
 
-        if (waitForView(bySelector, index) == null) {
+        if (waitForView(bySelector, index, timeout) == null) {
             throw UiObjectNotFoundException("$uiSelector")
         }
     }
@@ -355,7 +356,7 @@ class Automator private constructor() {
         return notifications
     }
 
-    fun tapOnNotification(index: Int) {
+    fun tapOnNotification(index: Int, timeout: Long? = null) {
         Logger.d("tapOnNotification($index)")
 
         try {
@@ -363,6 +364,10 @@ class Automator private constructor() {
                 resourceId = "android:id/status_bar_latest_event_content",
                 instance = index.toLong()
             )
+            val selector = query.toBySelector()
+            if (waitForView(selector, index, timeout) == null) {
+                throw UiObjectNotFoundException("$selector")
+            }
             val obj = uiDevice.findObject(query.toUiSelector())
             obj.click()
         } catch (err: UiObjectNotFoundException) {
@@ -372,9 +377,12 @@ class Automator private constructor() {
         delay()
     }
 
-    fun tapOnNotification(selector: UiSelector) {
+    fun tapOnNotification(selector: UiSelector, bySelector: BySelector, timeout: Long? = null) {
         Logger.d("tapOnNotification()")
 
+        if (waitForView(bySelector, 0, timeout) == null) {
+            throw UiObjectNotFoundException("$bySelector")
+        }
         val obj = uiDevice.findObject(selector)
         obj.click()
 
@@ -474,9 +482,9 @@ class Automator private constructor() {
     /**
      * Returns true if [bySelector] found a view at [index] within [timeoutMillis], false otherwise.
      */
-    private fun waitForView(bySelector: BySelector, index: Int): UiObject2? {
+    private fun waitForView(bySelector: BySelector, index: Int, timeout: Long? = null): UiObject2? {
         val startTime = System.currentTimeMillis()
-        while (System.currentTimeMillis() - startTime < timeoutMillis) {
+        while (System.currentTimeMillis() - startTime < (timeout ?: timeoutMillis)) {
             val objects = uiDevice.findObjects(bySelector)
             if (objects.size > index && objects[index] != null) {
                 return objects[index]