+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/Contracts.kt
@@ -174,21 +174,30 @@ class Contracts {
 
   data class TapRequest (
     val selector: Selector,
-    val appId: String
-  )
+    val appId: String,
+    val timeoutMillis: Long? = null
+  ){
+    fun hasTimeoutMillis(): Boolean {
+      return timeoutMillis != null
+    }
+  }
 
   data class EnterTextRequest (
     val data: String,
     val appId: String,
     val index: Long? = null,
     val selector: Selector? = null,
-    val keyboardBehavior: KeyboardBehavior
+    val keyboardBehavior: KeyboardBehavior,
+    val timeoutMillis: Long? = null
   ){
     fun hasIndex(): Boolean {
       return index != null
     }
     fun hasSelector(): Boolean {
       return selector != null
+    }
+    fun hasTimeoutMillis(): Boolean {
+      return timeoutMillis != null
     }
   }
 
@@ -203,8 +212,13 @@ class Contracts {
 
   data class WaitUntilVisibleRequest (
     val selector: Selector,
-    val appId: String
-  )
+    val appId: String,
+    val timeoutMillis: Long? = null
+  ){
+    fun hasTimeoutMillis(): Boolean {
+      return timeoutMillis != null
+    }
+  }
 
   data class DarkModeRequest (
     val appId: String
@@ -234,13 +248,17 @@ class Contracts {
 
   data class TapOnNotificationRequest (
     val index: Long? = null,
-    val selector: Selector? = null
+    val selector: Selector? = null,
+    val timeoutMillis: Long? = null
   ){
     fun hasIndex(): Boolean {
       return index != null
     }
     fun hasSelector(): Boolean {
       return selector != null
+    }
+    fun hasTimeoutMillis(): Boolean {
+      return timeoutMillis != null
     }
   }